layout-analysis/requirements.txt
absl-py==1.4.0
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.6
cloudpickle==2.2.1
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@4aca4bdaa9ad48b8e91d7520e0d0815bb8ca0fb1
fastapi==0.101.0
filelock==3.9.0
fonttools==4.42.0
future==0.18.3
fvcore==0.1.3.post20210317
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
h11==0.14.0
httptools==0.6.0
idna==3.4
iopath==0.1.10
Jinja2==3.1.2
kiwisolver==1.4.4
layoutparser==0.3.4
Markdown==3.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==9.3.0
portalocker==2.7.0
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydot==1.4.2
pyparsing==3.0.9
pypdfium2==4.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tqdm==4.66.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
yacs==0.1.8